+++ Projects/MerchantsGuide/RomanNumbersConversion/Domain/Metal.cs
@@ -9,6 +9,6 @@ namespace RomanNumbersConversion.Domain
     public class Metal
     {
         public string MetalName { get; set; }
-        public long MetalValue { get; set; }
+        public decimal MetalValue { get; set; }
     }
 }
